# js-api/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Apisme API",
    "version": "1.0.0",
    "description": "Documentação automática com Swagger Autogen"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "Users",
      "description": "Operações relacionadas a usuários"
    },
    {
      "name": "Products",
      "description": "Operações relacionadas a produtos"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/users/": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "Retorna todos os usuários",
        "responses": {
          "200": {
            "description": "Lista de usuários",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "name": {
                    "type": "string",
                    "example": "Alice"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "description": "Adiciona um novo usuário",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Dados do usuário",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Carlos"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "description": "Substitui todos os usuários",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Nova lista de usuários",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "name": {
                    "type": "string",
                    "example": "Alice"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "description": "Atualiza um usuário específico",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "ID e novos dados do usuário",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "name": {
                  "type": "string",
                  "example": "Alice Atualizada"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "description": "Remove um usuário pelo ID",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "ID do usuário",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "options": {
        "tags": [
          "Users"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "head": {
        "tags": [
          "Users"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/products/": {
      "get": {
        "tags": [
          "Products"
        ],
        "description": "Retorna todos os produtos",
        "responses": {
          "200": {
            "description": "Lista de produtos",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "name": {
                    "type": "string",
                    "example": "Notebook"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "description": "Adiciona um novo produto",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Dados do produto",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Teclado"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "description": "Substitui todos os produtos",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Nova lista de produtos",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "name": {
                    "type": "string",
                    "example": "Notebook"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Products"
        ],
        "description": "Atualiza um produto específico",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "ID e novos dados do produto",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "name": {
                  "type": "string",
                  "example": "Mouse Atualizado"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "description": "Remove um produto pelo ID",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "ID do produto",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "options": {
        "tags": [
          "Products"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "head": {
        "tags": [
          "Products"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    }
  }
}